refactor(PrefChart.vue, chartDataOperation.ts): チャート用の処理をTSファイルに分離。コンポーネント内で呼出しのみに変更
